fix: More defensive code when trying to drop tables

+++ migration/2_mnemo_upgrade_autoincrement.php
@@ -20,9 +20,8 @@ class MnemoUpgradeAutoIncrement extends Horde_Db_Migration_Base
     public function up()
     {
         $this->changeColumn('mnemo_shares', 'share_id', 'autoincrementKey');
-        try {
+        if (in_array('mnemo_shares_seq', $this->tables())) {
             $this->dropTable('mnemo_shares_seq');
-        } catch (Horde_Db_Exception $e) {
         }
     }
